requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai-whisper
ffmpeg-python
python-dotenv
google-generativeai
